优化

+++ bll/src/main/java/com/zjy/bll/dao/PermissionDao.java
@@ -6,13 +6,12 @@ import com.zjy.entities.Permission;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 @Repository
 public interface PermissionDao extends BaseDao<Permission> {
-    Map<String, BigDecimal> queryRepeatCount(@Param("permissionId") String permissionId, @Param("code") String code);
+    Map<String, Integer> queryRepeatCount(@Param("permissionId") String permissionId, @Param("code") String code);
 
     List<PermissionVo> queryAllPermissionList();
 }
